fix: deduplicate command bar into shared CheckSyncedCommandBar
